Update r136/ydb864 test to use white box test where needed
This commit updates the `r136/ydb864` test to use the new white box test
`WBTEST_YDB_STATICPID` for the `-noonline` portion of the test. This
addresses a test failure where the test could unexpectedly succeed for
the "DEFAULT" region with a path of 234 characters if the pid was 9999
or less. The white box test ensures that the process ID part of the temp
file name is always 99999 which ensures that the length of the temp file
will be as expected. Since white box test cases only work on dbg builds,
this commit also disables the `-noonline` part of the test on pro builds
and updates the reference file accordingly.

+++ r136/u_inref/ydb864.csh
@@ -51,26 +51,33 @@ if ("$firstbackup" =~ "*mumps.dat*") then
 endif
 rm -rf test
 
-@ cnt = 220
+if ("dbg" == "$tst_image") then
+	@ cnt = 220
 
+	setenv gtm_white_box_test_case_enable   1
+	setenv gtm_white_box_test_case_number   203      # WBTEST_YDB_STATICPID
 
-$echoline
-echo "# Starting the -noonline MUPIP BACKUPs. For these, we"
-echo "# generate directory paths of 220 to 266 characters,"
-echo "# expecting lengths of 220 to 233 to succeed and longer"
-echo "# lengths to result in errors. Lengths of 234 or larger"
-echo "# fail due to the space needed for the temporary file which"
-echo "# would otherwise overflow a buffer in sr_port/mubfilcpy.c"
+	$echoline
+	echo "# Starting the -noonline MUPIP BACKUPs. For these, we"
+	echo "# generate directory paths of 220 to 266 characters,"
+	echo "# expecting lengths of 220 to 233 to succeed and longer"
+	echo "# lengths to result in errors. Lengths of 234 or larger"
+	echo "# fail due to the space needed for the temporary file which"
+	echo "# would otherwise overflow a buffer in sr_port/mubfilcpy.c"
 
-while ($cnt < 266)
-	set dirname = `$gtm_dist/mumps -run gen $cnt`
-	rm -rf $dirname
-	mkdir $dirname
-	echo "####### Testing backup directory length [$cnt] with -noline #########"
-	$gtm_dist/mupip backup -noonline "*" $dirname
-	rm -rf $dirname
-	@ cnt = $cnt + 1
-end
+	while ($cnt < 266)
+		set dirname = `$gtm_dist/mumps -run gen $cnt`
+		rm -rf $dirname
+		mkdir $dirname
+		echo "####### Testing backup directory length [$cnt] with -noline #########"
+		$gtm_dist/mupip backup -noonline "*" $dirname
+		rm -rf $dirname
+		@ cnt = $cnt + 1
+	end
+
+	unsetenv WBTEST_YDB_STATICPID
+	unsetenv gtm_white_box_test_case_number
+endif
 
 $echoline
 echo "# Starting the online MUPIP BACKUPs. For these, we generate"
